py-simputils-config-38 Circular dependency if `ConfigStore` used without importing `ConfigHub`
* Adjusted tests structure to replicate an issue
